move objectAssignTemporalPurityCheck into its own file
